+++ test/dummy/config/routes.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  mount Testimonials::Engine => "/testimonials"
+  mount_testimonials at: "/testimonials"
   get "sample", to: "sample#show"
   post "sample/celebrate", to: "sample#celebrate"
 end